+++ .gitignore
@@ -144,3 +144,8 @@ docs/auto_examples/
 *-bem.fif
 *-fwd.fif
 *-src.fif
+
+# data files generated by our scripts that are too big for git
+data/paper/*.netcdf
+data/paper/*.edf
+
